requirements.txt
fastapi
uvicorn
pymupdf
pdfplumber
pytesseract
Pillow
opencv-python
sentence-transformers
faiss-cpu
numpy
python-dotenv
google-genai
pytest